+++ ProjectFactory/src/main/kotlin/net/milosvasic/factory/configuration/Configuration.kt
@@ -2,7 +2,9 @@ package net.milosvasic.factory.configuration
 
 object Configuration {
 
-    val repo = "http://repo.milosvasic.net/releases"
     val groot = Groot
+    val gradleVersion = "3.4.1"
+    val repo = "http://repo.milosvasic.net/releases"
+    val distributions = "http://milosvasic.net/distributions"
 
 }