+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Version 1.0.0 Alpha 2
 
-- Downloading local Gradle version.
+- Downloading local Gradle version
+- Initializing Gradle Wrapper
+- Building generated project
 
 # Version 1.0.0 Alpha 1
 
-- First working version.
+- First working version